requirements.txt
django==3.0.7
mysqlclient==1.4.6
pandas==1.0.4
numpy==1.18.5
celery[redis]==4.4.7
django-environ==0.4.5
Jinja2==2.11.2
dj_database_url==0.5.0
django-redis==4.12.1
sentry-sdk==0.19.5
chardet==3.0.4
idna==2.10
apscheduler==3.6.3
python-binance==1.0.12
coinmarketcap==5.0.3
django-environ==0.4.5
aiohttp==3.6.2
pycrypto==2.6.1
python-socketio[asyncio_client]==5.1.0
psutil==5.8.0
pyjwt==2.2.0